refactor: Controller 클래스 분리
